tests/test_requirements.txt
numpy
pandas
ipykernel
requests
matplotlib
pytest
pytest-cov
pyyaml